requirements.txt
pyautogui
opencv-python
pytesseract
keyboard
pywin32
numpy
pillow
